src/main/java/br/com/alura/escola/aplicacao/indicacao/aluno/matricular/MatricularAluno.java
package br.com.alura.escola.aplicacao.indicacao.aluno.matricular;

import br.com.alura.escola.dominio.aluno.Aluno;
import br.com.alura.escola.dominio.aluno.RepositorioDeAlunos;

public class MatricularAluno {

    private final RepositorioDeAlunos repositorio;

    public MatricularAluno(RepositorioDeAlunos repositorio) {
        this.repositorio = repositorio;
    }

    public void executa(MatricularALunoDto dados){
        Aluno novo = dados.criarALuno();
        repositorio.matricular(novo);
    }
}
